cloud storage download fix, consent flow update,

+++ lib/app/data/services/local_db.dart
@@ -1,23 +1,27 @@
 import 'dart:developer';
 import 'dart:io';
 
-import 'package:flutter/foundation.dart';
 import 'package:get/get.dart';
 import 'package:hive/hive.dart';
 import 'package:qualnote/app/data/models/project.dart';
 import 'package:qualnote/app/modules/home/controllers/home_controller.dart';
+import 'package:qualnote/app/utils/note_type.dart';
 
 class HiveDb extends GetxController {
   late Box<Project> projectsBox;
+  RxBool inProgress = false.obs;
 
   Future<void> deleteProjectLocaly(String projectId) async {
     var project = await getProject(projectId);
-    Get.find<HomeController>().projects.remove(project);
-    //uploads all notes to cloud storage
+    Get.find<HomeController>().localProjects.remove(project);
+
     for (var element in project!.notes!) {
+      if (element.type == NoteType.text.toString()) {
+        continue;
+      }
       await File(element.path!).delete();
     }
-    //upload main route recordings
+
     if (project.type! == 'RecordingType.video') {
       for (var path in project.routeVideos!) {
         await File(path).delete();
@@ -29,18 +33,13 @@ class HiveDb extends GetxController {
       }
     }
     projectsBox.delete(projectId);
-
     log('Project deleted from local storage');
   }
 
   Future<void> saveProject(Project project) async {
-    final homeController = Get.find<HomeController>();
-    await projectsBox.put(project.id, project);
-    var newProject = await getProject(project.id!);
-    homeController.projects.add(newProject!);
-    if (kDebugMode) {
-      print(project.toJson());
-    }
+    inProgress.value = true;
+    await projectsBox.put(project.id!, project);
+
     log('Route Saved');
   }
